netlink: drop support for Go 1.11
