--- rawconn_gteq_1.12.go
@@ -1,12 +1,0 @@
-//+build go1.12
-
-package netlink
-
-import (
-	"os"
-	"syscall"
-)
-
-func newRawConn(fd *os.File) (syscall.RawConn, error) {
-	return fd.SyscallConn()
-}
